experiments/fl-iot-ids-v1/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fl-iot-ids-v1"
version = "0.1.0"
description = "Federated Learning V1 for IoT IDS with Flower and PyTorch"
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.20.0",
    "torch>=2.2",
    "numpy>=1.26",
    "pandas>=2.2",
    "scikit-learn>=1.4",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "matplotlib>=3.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.flwr.app]
publisher = "saif-boukhatem"

[tool.flwr.app.components]
serverapp = "src.fl.server_app:app"
clientapp = "src.fl.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 3
learning-rate = 0.001
local-epochs = 1
batch-size = 256

[tool.flwr.federations]
default = "local-simulation"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 3